chore: organize call related stateful benchmark  - create test_call_operation file  - move test_call_value_existing and test_call_value_new_account  - rename test name
