+++ tests/benchmark/stateful/bloatnet/test_call_operation.py
@@ -1,0 +1,213 @@
+"""Benchmark call operations with value transfer on target accounts."""
+
+import pytest
+from execution_testing import (
+    Account,
+    Address,
+    Alloc,
+    BenchmarkTestFiller,
+    Block,
+    Bytecode,
+    Conditional,
+    Create2PreimageLayout,
+    Fork,
+    Hash,
+    IteratingBytecode,
+    Op,
+    While,
+    keccak256,
+)
+
+from tests.benchmark.stateful.helpers import (
+    DECREMENT_COUNTER_CONDITION,
+    build_benchmark_txs,
+)
+
+REFERENCE_SPEC_GIT_PATH = "DUMMY/bloatnet.md"
+REFERENCE_SPEC_VERSION = "1.0"
+
+
+@pytest.mark.stub_parametrize("factory_stub", "bloatnet_factory_")
+def test_call_value_to_existing(
+    benchmark_test: BenchmarkTestFiller,
+    pre: Alloc,
+    fork: Fork,
+    gas_benchmark_value: int,
+    tx_gas_limit: int,
+    factory_stub: str,
+) -> None:
+    """Benchmark CALL with value transfer to cold existing contracts."""
+    factory_address = pre.deploy_contract(
+        code=Bytecode(),
+        stub=factory_stub,
+    )
+
+    # Contract Construction
+    setup = Bytecode()
+
+    setup += Conditional(
+        condition=Op.STATICCALL(
+            gas=Op.GAS,
+            address=factory_address,
+            args_offset=0,
+            args_size=0,
+            ret_offset=96,
+            ret_size=64,
+            # gas accounting
+            address_warm=False,
+            old_memory_size=0,
+            new_memory_size=160,
+        ),
+        if_false=Op.INVALID,
+    )
+
+    create2_preimage = Create2PreimageLayout(
+        factory_address=factory_address,
+        salt=Op.CALLDATALOAD(32),
+        init_code_hash=Op.MLOAD(128),
+        old_memory_size=160,
+    )
+
+    setup += create2_preimage
+    setup += Op.CALLDATALOAD(0)  # [num_contract]
+
+    # CALL with value=1 to factory contracts.
+    # The address is computed inline via SHA3, avoiding DUP depth issues.
+    # gas=1: subcall gets 1 + 2300 stipend, still not enough for 24KB
+    # bytecode → subcall fails, but cold + value gas costs are charged.
+    call_value_op = Op.POP(
+        Op.CALL(
+            gas=1,
+            address=create2_preimage.address_op(),
+            value=1,
+            args_offset=0,
+            args_size=0,
+            ret_offset=0,
+            ret_size=0,
+            # gas accounting
+            value_transfer=True,
+        )
+    )
+
+    loop = While(
+        body=(call_value_op + create2_preimage.increment_salt_op()),
+        condition=DECREMENT_COUNTER_CONDITION,
+    )
+
+    # Contract Deployment
+    code = setup + loop
+    attack_contract_address = pre.deploy_contract(code=code)
+
+    # Gas Accounting
+    txs, total_gas_consumed = build_benchmark_txs(
+        pre=pre,
+        fork=fork,
+        gas_benchmark_value=gas_benchmark_value,
+        tx_gas_limit=tx_gas_limit,
+        attack_contract_address=attack_contract_address,
+        setup_cost=setup.gas_cost(fork),
+        iteration_cost=loop.gas_cost(fork),
+    )
+
+    benchmark_test(
+        pre=pre,
+        blocks=[Block(txs=txs)],
+        expected_benchmark_gas_used=total_gas_consumed,
+        skip_gas_used_validation=True,
+    )
+
+
+def test_call_value_to_empty(
+    benchmark_test: BenchmarkTestFiller,
+    pre: Alloc,
+    fork: Fork,
+    gas_benchmark_value: int,
+) -> None:
+    """Benchmark CALL with value transfer to non-existent accounts."""
+    # Memory layout: MEM[0..31] = counter (incremented each iteration)
+    setup = (
+        Op.MSTORE(
+            0,
+            Op.CALLDATALOAD(32),  # salt_offset (starting counter)
+            # gas accounting
+            old_memory_size=0,
+            new_memory_size=32,
+        )
+        + Op.CALLDATALOAD(0)  # [num_calls]
+    )
+
+    # CALL with value=1 to keccak256-derived addresses.
+    # gas=0: subcall gets 0 + 2300 stipend. No code at target → succeeds.
+    # Value is transferred, new account is created in trie.
+    call_value_op = Op.POP(
+        Op.CALL(
+            gas=0,
+            address=Op.SHA3(0, 32, data_size=32),
+            value=1,
+            args_offset=0,
+            args_size=0,
+            ret_offset=0,
+            ret_size=0,
+            # gas accounting
+            value_transfer=True,
+            account_new=True,
+        )
+    )
+
+    # Increment counter in memory for next address
+    increment_counter = Op.MSTORE(0, Op.ADD(Op.MLOAD(0), 1))
+
+    loop = While(
+        body=(call_value_op + increment_counter),
+        condition=DECREMENT_COUNTER_CONDITION,
+    )
+
+    # Contract Deployment — needs balance for value transfers (1 wei each)
+    code = IteratingBytecode(
+        setup=setup,
+        iterating=loop,
+    )
+
+    initial_balance = 10**9
+    attack_contract_address = pre.deploy_contract(
+        code=code,
+        balance=initial_balance,
+    )
+
+    def calldata_builder(iteration_count: int, start_iteration: int) -> bytes:
+        return bytes(Hash(iteration_count) + Hash(start_iteration))
+
+    txs = list(
+        code.transactions_by_gas_limit(
+            fork=fork,
+            gas_limit=gas_benchmark_value,
+            sender=pre.fund_eoa(),
+            to=attack_contract_address,
+            calldata=calldata_builder,
+        )
+    )
+
+    total_iterations = sum(int.from_bytes(tx.data[:32], "big") for tx in txs)
+
+    def new_account_address(counter: int) -> Address:
+        return Address(bytes(keccak256(counter.to_bytes(32, "big")))[12:])
+
+    post = {
+        new_account_address(counter): Account(balance=1)
+        for counter in range(total_iterations)
+    }
+    post[attack_contract_address] = Account(
+        balance=initial_balance - total_iterations
+    )
+
+    expected_gas_used = (
+        sum(tx.gas_cost for tx in txs)
+        - fork.gas_costs().CALL_STIPEND * total_iterations
+    )
+
+    benchmark_test(
+        pre=pre,
+        post=post,
+        blocks=[Block(txs=txs)],
+        expected_benchmark_gas_used=expected_gas_used,
+    )
